Implement Solution for Day 02 (#16)
* feat: implement solution for Day 02 riddle 1

Add the `SolveDay02RiddleUseCase` class to handle the logic for solving the Day 02 riddle. Introduce `ProductRange` and `ProductRanges` classes for managing ranges from input files. Update the main application to include the new riddle solution and adjust the greeting message.

* feat: implement solution for Day 02 riddle 2

Add logic to generate numbers with repeating digit blocks and check them against defined ranges. This includes creating a set of valid numbers and returning the first one that fits within the specified ranges.

* chore: update group ID for project configuration

Change the group ID in the Gradle build file to reflect the new namespace for the project.

* style: run spotless